+++ core_client/base/api/v3_fs_put_file.py
@@ -8,7 +8,7 @@ from ..models import Error
 @validate_arguments()
 def _build_request(
     client: Client,
-    name: str,
+    storage: str,
     path: str,
     data: bytes,
     retries: int = None,
@@ -20,7 +20,7 @@ def _build_request(
         timeout = client.timeout
     return {
         "method": "put",
-        "url": f"{client.base_url}/api/v3/fs/{name}/{path}",
+        "url": f"{client.base_url}/api/v3/fs/{storage}/{path}",
         "headers": client.headers,
         "timeout": timeout,
         "content": data,